+++ packages/frontend/@n8n/i18n/src/locales/en.json
@@ -1497,6 +1497,7 @@
 	"nodeView.couldntImportWorkflow": "Could not import workflow",
 	"nodeView.couldntLoadWorkflow.invalidWorkflowObject": "Invalid workflow object",
 	"nodeView.deletesTheCurrentExecutionData": "Deletes the current execution data",
+	"nodeView.focusPanel.noExecutionData": "Execute Node for autocomplete",
 	"nodeView.focusPanel.noParameters.title": "Show a node parameter here, to iterate easily",
 	"nodeView.focusPanel.noParameters.subtitle": "For example, keep your prompt always visible so you can run the workflow while tweaking it",
 	"nodeView.focusPanel.missingParameter": "This parameter is no longer visible on the node. A related parameter was likely changed, removing this one.",
